[CMD] CTTY: Improve opening the STD TTY handles.
Opening the StdOut handle in both write and read access
allows one to retrieve any possible result/status from an
ANSI escape query (for example, asking where the cursor is
or the current size of the screen).

Also, make the StdErr handle just inherit from StdOut with
the same access.