[SETUPLIB] In PrepareCopyInfFile(), explicitly use AddSectionToCopyQueueCab() to retrieve the files from the CAB INF.
Otherwise if one continued to use and rely on AddSectionToCopyQueue()
to do this job instead, bad things could happen like, re-enumerating
twice or more (and adding them to the file queue) the same files.
This is because in the new code AddSectionToCopyQueue() can be called
many times to deal with different SIF file sections. The old code then
called AddSectionToCopyQueueCab() in turn, but on the same list...
Now the steps performed are clear, and everything works as intended.

Additional changes:
- Shut a lot of DPRINTs up.
- Fix the code that builds the name of the CAB INF file from the CAB
  file name: it should first find for the filename in the full path,
  and then find the file extension (if any) and change it to ".inf".