[REACTOS] Add helpers and modify ConvertNtPathToWin32Path() to make it use a cache of NT/Win32 path mappings.
This increases performance for each time the SETUPLIB calls, with NT paths,
Win32 SetupAPI functions which of course accept only Win32 paths.
Handle also the fact that a NT path to convert may start with
\Device\HarddiskX\PartitionY\..., which can be a symlink to
\Device\HarddiskVolumeN\... on some systems, and in that case, the Win32
path mapping should be done slightly differently (due to the symlink).

IN ADDITION: start to plug the update-registry code in.
BEWARE of the fact it currently contains parameters specific to my
testing machine!! (you need to find & fix them up if you wish to test
this code for yourself).
The code has also been tested on Windows and it works as expected.