Experimenting with RtlCreateUserThread...
Note that this somehow introduces a bug: Since we are creating thread
for a Win32 console app, this thread should also be registered wrt. CSR.
This is done automatically by kerenl32!Create(Remote)Thread, but not by
ntdll!RtlCreateUserThread (and thus it would be necessary to do it by
hand). The observable effect of this difference is that when a console
app is killed by directly closing its console window (thus emitting a
CTRL_CLOSE signal to the app via this thread creation), an LPC
connection error is also emitted, which does not happen if the thread
is also registered with CSR.