[NTOS:CONFIG] CmpOpenHiveFiles(): Do not return potential invalid file dispositions.
The hive and log files creation dispositions, that this function return
to its callers, are obtained from the ZwCreateFile() calls via the
IoStatusBlock's Information field. But they are actually returned only
once the file creation/opening + auxiliary initialization operations are
performed. However, the IoStatusBlock was recycled for these operations
as well, namely the ZwFsControlFile() calls, and therefore could be
overwritten before we use it for the file creation disposition.
Use now a separate IO status block for the ZwFs* calls so as to not
overwrite the other one.