[FREELDR:NTLDR] The Synthetic NT/ROS Loader At Last.
Finish to unify the standard loader with the Setup loader as 98% of these
are common functionality. Distinguish between the two only in few places
via `if (IsNTSetup)` or similar checks, and using helper functions for
the functionality that has to differ between the two.

This finishes the unification started by commit 0c902fc49 (r53510).

+ Reorder load steps to be closer to what Windows' NTLDR does.