[NTOS:KE] Remove KiInitModuleList hack.
It was introduced in commit 89b75ad8 (r53714) in order to work-around
the fact the PsLoadedModuleList is only initialized later on by
MiInitializeLoadedModuleList(), called from Mm(Arm)InitSystem().
The correct solution is to look at KeLoaderBlock->LoadOrderListHead
and let the debugger manage from where it gets the list of boot images
to load the symbols from.