[NTOS:KDBG] Isolate terminal and input-related routines from the rest of KDBG. (#5188)
[NTOS:KD:KDBG] Adapt the code to a form closer to what it needs to be when terminal/input functionality will be in a separate DLL.

The separate DLL, KDTERM.DLL, will be a KD transport DLL and as such
will export the usual functions. But in addition, it will also export a
symbol "KdTransport", which is a structure representing the controlling
terminal -- including few manipulation functions.

In addition, integrate the KdbDebugState flags (KDSERIAL and KDNOECHO)
into the terminal state: respectively, SerialInput and NoEcho bitfields.