** WIP ** More work to make KDTERM.DLL work on Windows.
In particular, the KD DLL is a boot-loaded module, that is relocated
by the kernel (see MiReloadBootLoadedDrivers) when the Memory Manager
initializes. This however happens after KdDebuggerInitialize0 has been
called and we have initialized some internal structures, and before
calling KdDebuggerInitialize1. Debug prints or other events may arise
between those two calls and just after the relocation. Hence, extra
care must be taken when initializing data.

Rewrite the KD IO providers initialization as the previous code was
incompatible with this state of affairs.