** WIP ** [FREELDR] Split up the menu code into actual menu and "screen" concepts.
The idea is to make the menu code (in uimenu and friends) just what
it should be: managing menus. The rest of the decorations (backdrop,
banner/title/header/footer/timeout/other text...) are external to the
menu, and are part of the displayed "screen". (See e.g. the minitui
NTLDR-like FreeLdr interface.)

This would allow, in the future, to support multiple menus (sub-menus)
managed from within the same screen. Currently, we can only support one
menu per screen.