** Super WIP ** [CONSOLE] Turn it as a hybrid CPL/Shell Property Pages extension for shortcuts
Instead of doing like in Windows and having shell32 duplicating the
functionality of console.dll for showing console-specific property pages
for shell links to console apps, use instead the console.dll by allowing
it to act as a Shell Property Pages extension for shortcuts, that
displays the pages for console apps.

** WIP ** CONSOLE.CPL Modify the CLSID guid to match the one used in microsoft/terminal console dll

Indeed, since Windows 10/11 they don't duplicate the console properties
pages & code for shortcuts in shell32, but use the existing console CPL
also as a shell property sheet handler.
And we do the same.